backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.1
pydantic-settings==2.2.1
SQLAlchemy==2.0.27
pymysql==1.1.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
redis==5.0.1
celery==5.3.6